# Handle non-Hash to Hash transformation in `before(:key_coercer)`